requirements-dev.txt
-r requirements.txt
WebOb==1.2.3
WebTest==2.0.8
beautifulsoup4==4.3.1
coverage==3.6
django-debug-toolbar==0.9.4
django-discover-runner==1.0
django-webtest==1.7.5
factory-boy==2.1.2
flake8==2.0
mccabe==0.2.1
pep8==1.4.6
pep8-naming==0.2.1
pyflakes==0.7.3
waitress==0.8.7
